\name{ExampleBSseq}
\alias{ExampleBSseq}
\title{BSseq example dataset}
\description{
A simulated example dataset of BSseq for MACAU2.
}
\format{
  Contains the following objects:
  \describe{
  	\item{mcount}{a matrix containing the read counts for 100 sites.}
    \item{predictor}{a vector (size=200) of observations on a continuous variable.}
    \item{relatednessmatrix}{a genetic relationship matrix for 200 individuals.}
    \item{totalcount}{a matrix containing the total read counts for 100 sites.}
  }
}


\keyword{example}
\keyword{BSseq}