\name{ExampleRNAseq}
\alias{ExampleRNAseq}
\title{RNAseq example dataset}
\description{
A simulated example dataset of RNAseq for MACAU2.
}
\format{
  Contains the following objects:
  \describe{
  	\item{count}{a matrix containing the read counts for 100 genes.}
    \item{predictor}{a vector (size=200) of observations on a continuous variable.}
    \item{relatednessmatrix}{a genetic relationship matrix for 200 individuals.}
	\item{totalcount}{a vector (size=200) with one read depth number for each individual.}
  }
}

\keyword{example}
\keyword{RNAseq}